# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: HireMind AI API Spec
  description: Spécification des contrats d'API pour la plateforme SaaS de recrutement HireMind AI. Ce document définit l'interface d'intégration entre le Client Mobile (Flutter), le Portail Web (Next.js 15) et le Core Backend (NestJS).
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Serveur de développement local NestJS
paths:
...
  /jobs:
    post:
      summary: Créer une offre d'emploi assistée par IA (Recruteur)
      description: Le recruteur soumet un prompt rapide et l'IA génère la fiche de poste complète (missions, soft skills, salaire marché).
      tags:
        - Offres d'emploi
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - prompt
              properties:
                prompt:
                  type: string
                  example: "Développeur Flutter Senior, 5 ans d'expérience, télétravail partiel, Paris"
      responses:
        '201':
          description: Fiche de poste générée et enregistrée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "job_018273"
                  title:
                    type: string
                    example: "Développeur Mobile Flutter Senior (H/F)"
                  description:
                    type: string
                    example: "Nous recherchons un développeur senior pour piloter notre transition vers Flutter..."
                  requirements:
                    type: object
                    properties:
                      technical:
                        type: array
                        items:
                          type: string
                        example: ["Flutter", "Dart", "CI/CD", "Clean Architecture"]
                      softSkills:
                        type: array
                        items:
                          type: string
                        example: ["Autonomie", "Mentorat", "Rigueur"]
                  salaryRange:
                    type: string
                    example: "55k€ - 65k€"

    get:
      summary: Lister les offres d'emploi disponibles
      tags:
        - Offres d'emploi
      responses:
        '200':
          description: Liste des offres d'emploi.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    title:
                      type: string
                    description:
                      type: string
  /jobs/{id}/match:
    post:
      summary: Calculer le score de matching sémantique d'un candidat
      description: Calcule la distance cosinus entre le CV du candidat et l'offre d'emploi à l'aide de Sentence Embeddings et Qdrant.
      tags:
        - Offres d'emploi
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "job_018273"
      responses:
        '200':
          description: Score de matching calculé avec détails.
          content:
            application/json:
              schema:
                type: object
                properties:
                  matchingScore:
                    type: number
                    format: float
                    description: Pourcentage de matching global.
                    example: 78.5
                  matchingBreakdown:
                    type: object
                    properties:
                      technicalMatch:
                        type: number
                        example: 90
                      experienceMatch:
                        type: number
                        example: 70
                      softSkillsMatch:
                        type: number
                        example: 75
                  keywordsMatched:
                    type: array
                    items:
                      type: string
                    example: ["Flutter", "Clean Architecture", "Dart"]

  # --- IA ADAPTIVE INTERVIEW ---
  /interviews/start:
    post:
      summary: Lancer une session d'entretien IA
      description: Initialise une session d'entretien adaptatif liée à une offre d'emploi spécifique.
      tags:
        - Entretien IA Adaptatif
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jobId
              properties:
                jobId:
                  type: string
                  example: "job_018273"
      responses:
        '201':
          description: Session d'entretien démarrée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                    example: "sess_550e8400-e29b-41d4-a716-446655440000"
                  firstQuestion:
                    type: string
                    example: "Bonjour ! Pouvez-vous me présenter vos expériences marquantes en développement d'applications avec Flutter ?"
  /interviews/{sessionId}/message:
    post:
      summary: Envoyer une réponse textuelle et obtenir la question suivante
      description: Soumet la réponse textuelle du candidat à l'arbre décisionnel IA et récupère la question suivante adaptée au niveau.
      tags:
        - Entretien IA Adaptatif
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          example: "sess_550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  example: "J'ai développé une application de e-commerce avec Flutter et BLoC en Clean Architecture."
      responses:
        '200':
          description: Réponse traitée et prochaine question renvoyée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  nextQuestion:
                    type: string
                    example: "Très intéressant. Pourquoi avez-vous privilégié BLoC par rapport à d'autres solutions de gestion d'état comme Riverpod pour ce projet ?"
                  isFinished:
                    type: boolean
                    description: Indique si l'entretien est terminé.
                    example: false
  /interviews/{sessionId}/audio:
    post:
      summary: Envoyer une réponse vocale et obtenir la question suivante
      description: Upload un chunk audio de la réponse du candidat (Whisper STT sur le Backend) et retourne la transcription et la prochaine question.
      tags:
        - Entretien IA Adaptatif
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - audioFile
              properties:
                audioFile:
                  type: string
                  format: binary
                  description: Fichier audio encodé (ex. m4a, wav).
      responses:
        '200':
          description: Audio transcrit et prochaine question générée.
          content:
            application/json:
              schema:
                type: object
                properties:
                  transcription:
                    type: string
                    example: "J'utilise principalement GoRouter pour gérer les routes de manière déclarative."
                  nextQuestion:
                    type: string
                    example: "Comment gérez-vous la redirection de routes pour les utilisateurs non authentifiés avec GoRouter ?"
                  nextQuestionAudioUrl:
                    type: string
                    format: uri
                    description: URL du flux audio de la question généré par ElevenLabs TTS.
                    example: "https://api.hiremind.ai/v1/assets/audio/q_98273.mp3"
                  isFinished:
                    type: boolean
                    example: false

  # --- SANDBOX EXECUTION (TESTS TECHNIQUES) ---
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT